+++ example/lib/pages/tables/tables.dart
@@ -3,6 +3,7 @@ import 'package:ng_admin/ng_admin.dart';
 import 'package:ng_admin_demo/components/how_to_use/how_to_use.dart';
 import 'table1.dart';
 import 'table2.dart';
+import 'table_with_filter.dart';
 
 @Component(
   selector: 'tables',
@@ -12,6 +13,7 @@ import 'table2.dart';
     ngAdminDirectives,
     Table1Component,
     Table2Component,
+    TableFilterComponent,
     HowToUseComponent
   ],
 )